+++ NEWS.md
@@ -1,6 +1,7 @@
 # motif 0.6.1
 
 * `lsp_add_stars` and `lsp_add_terra` now do not drop the `signature` list-column, but restructures it into many raster layers
+* Changes defaults in `lsp_signature` to `ordered = FALSE` and `repeated = FALSE` (both were `TRUE` before)
 
 # motif 0.6.0